reorganization, adapt text for docs site
