--- src/plot_label_attributes/Project.toml
@@ -1,7 +1,0 @@
-[deps]
-Genie = "c43c736e-a2d1-11e8-161f-af95117fbd1e"
-GenieFramework = "a59fdf5c-6bf0-4f5d-949c-a137c9e2f353"
-
-[compat]
-Genie = "5.0.0"
-GenieFramework = "1.0.0"
